src/requirements.txt
async-timeout==4.0.2
attrs==21.4.0
branca==0.5.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cycler==0.11.0
Deprecated==1.2.13
Fiona==1.8.21
Flask==2.1.2
folium==0.12.1.post1
fonttools==4.33.3
geopandas==0.10.2
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.3
MarkupSafe==2.1.1
matplotlib==3.5.2
munch==2.5.0
numpy==1.22.4
openrouteservice==2.3.3
packaging==21.3
paho-mqtt==1.6.1
pandas==1.4.2
Pillow==9.1.1
pymongo==4.1.1
pyparsing==3.0.9
pyproj==3.3.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.3
requests==2.28.0
scipy==1.8.1
seaborn==0.11.2
Shapely==1.8.2
six==1.16.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
